fix(watch): use Int64 for millisecond timestamps on ARM64_32
Apple Watch is ARM64_32 where Swift's `Int` is 32-bit. The B129 logging
instrumentation sprinkled `Int(Date().timeIntervalSince1970 * 1000)`
across eight call sites in the Watch target. Epoch-ms values in 2026
(~1.78e12) exceed Int32.max (~2.15e9), so Swift's strict
Double→Int overflow checking trapped with EXC_BREAKPOINT (code 0x1) on
whatever thread called the delegate — typically Thread 5 on a utility
QoS NSOperationQueue dispatched by WatchConnectivity. B130 crash log
register dump confirmed: x8 held 0x41e0000000000000 (IEEE-754 double
representation of exactly Int32.max + 1) as the overflow threshold.

Changes:
- All eight `Int(Date().timeIntervalSince1970 * 1000)` → `Int64(...)`
  in WatchSessionManager, AuthService, HomeView.
- pushTsMs cross-platform read: `as? Int ?? 0` → `as? Int64 ?? 0`.
  iPhone writes this value via NSNumber bridge; `as? Int` on the
  Watch (Int32) silently returned nil for values > Int32.max, making
  latencyMs always -1.
- Added a comment at the top of the WCSessionDelegate extension
  documenting the rule, so future devs don't "fix" Int64 back to Int.

Saved a memory (feedback_arm64_32_int_pitfall.md) covering the rule,
the symptom, and the grep patterns to spot it in future reviews.

B131 fix. Only the Watch target is affected — no API changes.

+++ apps/watch/KanjiLearnWatch/Services/AuthService.swift
@@ -76,7 +76,7 @@ final class AuthService {
 
     /// Returns a valid (non-expired) access token, refreshing autonomously if needed.
     func getAccessToken() async throws -> String {
-        let ts = Int(Date().timeIntervalSince1970 * 1000)
+        let ts = Int64(Date().timeIntervalSince1970 * 1000)
 
         guard let accessToken = load(key: KeychainKey.accessToken) else {
             print("[KL-Watch] \(ts) auth.getAccessToken result=missing")
@@ -103,7 +103,7 @@ final class AuthService {
 
     @discardableResult
     func refresh() async throws -> String {
-        let ts = Int(Date().timeIntervalSince1970 * 1000)
+        let ts = Int64(Date().timeIntervalSince1970 * 1000)
         guard let refreshToken = load(key: KeychainKey.refreshToken),
               let supabaseURL  = load(key: KeychainKey.supabaseURL) else {
             print("[KL-Watch] \(ts) auth.refresh result=no-refresh-token-or-url")
@@ -160,7 +160,7 @@ final class AuthService {
     // ── Sign out ──────────────────────────────────────────────────────────────
 
     func clear() {
-        let ts = Int(Date().timeIntervalSince1970 * 1000)
+        let ts = Int64(Date().timeIntervalSince1970 * 1000)
         print("[KL-Watch] \(ts) auth.clear called")
         delete(key: KeychainKey.accessToken)
         delete(key: KeychainKey.refreshToken)
